converted person to size_t

+++ numberphile/switch/switch_c/switch.c
@@ -1,9 +1,9 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-void print_grid(int person, char *buttons, size_t size){
+void print_grid(size_t person, char *buttons, size_t size){
     size_t n = size + 1;
-    printf("person: %d\n", person);
+    printf("person: %ld\n", person);
     for(int i=1; i<n; i++){
         printf("%d", buttons[i-1]);
         if (i % 10 == 0) printf("\n");
